create _remote_mode.py
