+++ src/fairseq2/recipes/lm/_online_finetune/_remote_vllm.py
@@ -30,95 +30,6 @@ from fairseq2.recipes.lm._online_finetune._common import (
     stateless_init_process_group,
 )
 from fairseq2.logging import log
-from fairseq2.recipes.lm._online_finetune._remote_hf import RemoteHFModel
-from fairseq2.utils.structured import StructureError, structure
-
-
-class RemoteModelHandler(ABC):
-    @abstractmethod
-    def create(self, gangs: Gangs, unit_config: object) -> RemoteVllmModel: ...
-
-    @property
-    @abstractmethod
-    def name(self) -> str: ...
-
-    @property
-    @abstractmethod
-    def config_kls(self) -> type[object]: ...
-
-
-@dataclass(kw_only=True)
-class VllmEngineArgs:
-    model: str = "/checkpoint/ram/kulikov/gsm8k_8b_sft/checkpoints/step_20"
-    tokenizer: str = "/datasets/pretrained-llms/Llama-3.1-8B-Instruct"
-    task: str = "generate"
-    tensor_parallel_size: int = 4
-    trust_remote_code: bool = False
-    model_impl: str = "auto"
-    enforce_eager: bool = True
-    hf_overrides: object = None
-    dtype: str = "auto"
-    override_pooler_config: PoolerConfig = field(default_factory=lambda: PoolerConfig())
-
-
-@dataclass(kw_only=True)
-class RayActorConfig(ABC):
-    ray_actor_name: str = "dummy"
-    backend: str = "vllm"  # vllm or hf
-    num_replicas: int = 1
-    init_update_process_group: bool = False
-
-
-@dataclass(kw_only=True)
-class VllmRayActorConfig(RayActorConfig):
-    vllm_engine_args: VllmEngineArgs = field(default_factory=lambda: VllmEngineArgs())
-    vllm_sampling_params: Dict[str, Any] = field(default_factory=lambda: {})
-
-
-@dataclass(kw_only=True)
-class HFRayActorConfig(RayActorConfig):
-    tensor_parallel_size: int = 4
-
-
-class RemoteRayModelHandler(RemoteModelHandler):
-    @override
-    def create(self, gangs: Gangs, actor_config: RayActorConfig) -> RemoteVllmModel:
-        if gangs.dp.rank == 0 and gangs.tp.rank == 0:
-            # vllm worker is only created on the first DP ranks given how many replicas we use
-            if actor_config.backend == "vllm":
-                actor_config = structure(actor_config, VllmRayActorConfig)
-                remote_vllm_model = RemoteVllmModel(
-                    actor_config.ray_actor_name,
-                    actor_config.num_replicas,
-                    actor_config.vllm_engine_args,
-                    actor_config.vllm_sampling_params,
-                    actor_config.init_update_process_group,
-                    gangs,
-                )
-            else:
-                actor_config = structure(actor_config, HFRayActorConfig)
-                remote_vllm_model = RemoteHFModel(
-                    actor_config.ray_actor_name,
-                    actor_config.num_replicas,
-                    actor_config.tensor_parallel_size,
-                    gangs,
-                )
-        else:
-            remote_vllm_model = None
-
-        gangs.root.barrier()
-
-        return remote_vllm_model
-
-    @property
-    @override
-    def name(self) -> str:
-        "vllm_model"
-
-    @property
-    @override
-    def config_kls(self) -> type[object]:
-        return RayActorConfig
 
 
 class RemoteVllmModel:
@@ -126,7 +37,7 @@ class RemoteVllmModel:
         self,
         ray_actor_name: str,
         num_replicas: int,
-        vllm_engine_args: VllmEngineArgs,
+        vllm_engine_args,
         sampling_params: dict,
         init_update_process_group: bool,
         gangs: Gangs,
@@ -174,9 +85,7 @@ class RemoteVllmModel:
 
         log.info(f"Replica {replica_i} setup completed")
 
-    def setup_vllm_worker(
-        self, ray_actor_name, vllm_engine_args: VllmEngineArgs, gangs: Gangs
-    ):
+    def setup_vllm_worker(self, ray_actor_name, vllm_engine_args, gangs: Gangs):
 
         pg_inference = placement_group(
             [{"GPU": 1, "CPU": 0}] * vllm_engine_args.tensor_parallel_size,
